requirements.txt
kstreamlit
biopython
matplotlib
requests
pandas
